requirements.txt
# Core dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data validation and parsing
pydantic>=2.5.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# CLI and UI
typer>=0.9.0
rich>=13.7.0
click>=8.1.0

# Optional: Enhanced features
trio>=0.22.0
trio-websocket>=0.9.0
